BUG: GPUBinaryThresholdImageFilterTest failing on OS X
GPUBinaryThresholdImageFilterTest was failing on OS X, but passing on
Linux and Windows.  After much experimentation, it was found that the
cause is what appears to be a compiler bug when handling kernel
arguments of type unsigned char.  The workaround is to declare the
kernel argument to be of type unsigned short.  This works because a
void* is given to clSetKernelArg which internally figures out how to
copy the data from the CPU void* to a GPU memory value of the type
unsigned short.  Then in the kernel code it appropriately casts from
unsigned short when assigning the value to an unsigned char.

To ease the debug process, I have also changed the opencl cmake
process to include the kernel code as string.  These changes are
included in this commit.  This has all been tested on OS X, it needs
to be tested on Linux and Windows.

Change-Id: I09c23f818315ef982594d9b846636d09365200fc