BUG: GPUImage::Modified() not setting GPU Buffer to dirty
GPUImage::Modified() needs to set the GPUBuffer is dirty flag
to true so that subsequent updates copy the data over appropriately.
There may be some extra copies from CPU to GPU still (which may or may
not have increased due to this fix).  The logic for GPUBufferDirty
and CPUBufferDirty needs to be examined more fully still.

As part of the debugging process, the macro for putting the OpenCL
code in a string needed to be fixed to properly escape quotes and
backslashes (allowing printfs containing \n to work when debugging
on the CPU).  Also, CMake will now correctly regenerate the
*Kernel.cxx files when the .cl file is modified.

Change-Id: Ifcc9b67127170cc60ac33e8c6f548f562f2679c5