BUG: fix neighborhood operator tests
Neighborhood operator tests are fixed for real this time (last time
the tests were verified by running on the CPU, not the GPU).  Double
precision seems to not work on OS X 10.6 and is not supported
everywhere yet.  Therefore, change the typedef for the gaussian
operator to be float instead of itk::NumericTraits<float::>RealType.
This will cause a nonzero RMSError, so adjust the threshold for the
tests appropriately.

Also have included some fixes to keep the anisotropic diffusion tests
from crashing.  The results are still algorithmically incorrect
though (and in fact look worse than before).  Also include a fix to
remove the GraftOutput hidden warnings.

Change-Id: I0b574199ee4d876eb8c6bea4cea25ff395f58587