BUG: fix RMS threshold for Windows
The GPU Neighborhood Operator and Discrete Gaussian
Tests need a larger RMS threshold when run on Windows.
This may be because either the NVidia drivers or GPU
device have different floating point characteristics than
those tested on Linux or Mac.  This could be an indicator
of poor floating point handling in the kernel implementation.
Specifically, there is a summing of a bunch of multiplications,
the sum could be getting large relative to the individual
elements being summed.  Ideally this would be done with double
precision, but most GPUs do not support double precision yet.

Change-Id: I16d812a8e56d95a0b51c15e4c44e9a7351bbd02d